requirements.txt
playwright
openpyxl
python-docx
docx2pdf
PyPDF2
spacy